requirements.txt
asgiref==3.9.1
cffi==1.17.1
cryptography==45.0.6
Django==5.2.5
django-cors-headers==4.7.0
django-encrypted-model-fields==0.6.5
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.1.1
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.9.0
